+++ scripts/templates/universal-react/server/config.js
@@ -1,6 +1,6 @@
 /**
  * This is the app config file. It's defined in the server, so you can access process.env.
- * In React, AppDataContext can be used to access this config in both the client and server.
+ * In React, useAppData can be used to access this config in both the client and server.
  */
 module.exports = () => ({
   //