+++ universal-react/index.js
@@ -1,3 +1,4 @@
 import AppDataContext from './AppDataContext';
+import useAppData from './useAppData';
 
-export { AppDataContext };
+export { AppDataContext, useAppData };