+++ universal-react/useAppData.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import AppDataContext from './AppDataContext';
+
+export default function useAppData() {
+  return React.useContext(AppDataContext);
+}